refactor(commands,llm): propagate GoalCommand errors, deduplicate test helpers, consolidate EmaTracker state
- GoalCommand::handle now propagates handle_goal errors via ? instead of
  silently converting them to display strings with unwrap_or_else (closes #3933)
- Shared test mocks (MockDebug, MockMessages, MockSession, make_ctx) extracted
  from 14 handler modules into handlers::test_helpers, removing ~1400 lines of
  duplicated boilerplate (closes #3926)
- EmaTracker state consolidated from two Arc<Mutex<>> fields into a single
  Arc<parking_lot::RwLock<EmaState>>, eliminating the TOCTOU window in
  maybe_reorder and enabling concurrent reads in snapshot (closes #3892)
