+++ crates/zeph-commands/src/handlers/goal.rs
@@ -51,7 +51,7 @@ impl CommandHandler<CommandContext<'_>> for GoalCommand {
         let span = tracing::info_span!("commands.goal.handle");
         Box::pin(
             async move {
-                let result = ctx.agent.handle_goal(args).await.unwrap_or_else(|e| e.0);
+                let result = ctx.agent.handle_goal(args).await?;
                 Ok(CommandOutput::Message(result))
             }
             .instrument(span),
@@ -62,81 +62,8 @@ impl CommandHandler<CommandContext<'_>> for GoalCommand {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::context::CommandContext;
+    use crate::handlers::test_helpers::{MockDebug, MockMessages, MockSession, make_ctx};
     use crate::sink::NullSink;
-    use crate::traits::debug::DebugAccess;
-    use crate::traits::messages::MessageAccess;
-    use crate::traits::session::SessionAccess;
-    use std::future::Future;
-    use std::pin::Pin;
-
-    struct MockDebug;
-    impl DebugAccess for MockDebug {
-        fn log_status(&self) -> String {
-            String::new()
-        }
-        fn read_log_tail<'a>(
-            &'a self,
-            _n: usize,
-        ) -> Pin<Box<dyn Future<Output = Option<String>> + Send + 'a>> {
-            Box::pin(async { None })
-        }
-        fn scrub(&self, text: &str) -> String {
-            text.to_owned()
-        }
-        fn dump_status(&self) -> Option<String> {
-            None
-        }
-        fn dump_format_name(&self) -> String {
-            String::new()
-        }
-        fn enable_dump(&mut self, _dir: &str) -> Result<String, CommandError> {
-            Ok(String::new())
-        }
-        fn set_dump_format(&mut self, _name: &str) -> Result<(), CommandError> {
-            Ok(())
-        }
-    }
-
-    struct MockMessages;
-    impl MessageAccess for MockMessages {
-        fn clear_history(&mut self) {}
-        fn queue_len(&self) -> usize {
-            0
-        }
-        fn drain_queue(&mut self) -> usize {
-            0
-        }
-        fn notify_queue_count<'a>(
-            &'a mut self,
-            _count: usize,
-        ) -> Pin<Box<dyn Future<Output = ()> + Send + 'a>> {
-            Box::pin(async {})
-        }
-    }
-
-    struct MockSession;
-    impl SessionAccess for MockSession {
-        fn supports_exit(&self) -> bool {
-            false
-        }
-    }
-
-    fn make_ctx<'a>(
-        sink: &'a mut NullSink,
-        debug: &'a mut MockDebug,
-        messages: &'a mut MockMessages,
-        session: &'a MockSession,
-        agent: &'a mut crate::NullAgent,
-    ) -> CommandContext<'a> {
-        CommandContext {
-            sink,
-            debug,
-            messages,
-            session: session as &dyn SessionAccess,
-            agent,
-        }
-    }
 
     #[test]
     fn goal_name_and_description() {
@@ -145,27 +72,27 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn goal_returns_message_when_agent_errors() {
-        // NullAgent default handle_goal returns Err, which is unwrapped to the error message.
+    async fn goal_propagates_error_from_agent() {
+        // NullAgent::handle_goal returns Err — the handler must propagate it.
         let mut sink = NullSink;
         let mut debug = MockDebug;
         let mut messages = MockMessages;
         let session = MockSession;
         let mut agent = crate::NullAgent;
         let mut ctx = make_ctx(&mut sink, &mut debug, &mut messages, &session, &mut agent);
-        let out = GoalCommand.handle(&mut ctx, "status").await.unwrap();
-        assert!(matches!(out, CommandOutput::Message(_)));
+        let result = GoalCommand.handle(&mut ctx, "status").await;
+        assert!(result.is_err());
     }
 
     #[tokio::test]
-    async fn goal_with_empty_args_returns_message() {
+    async fn goal_with_empty_args_propagates_error() {
         let mut sink = NullSink;
         let mut debug = MockDebug;
         let mut messages = MockMessages;
         let session = MockSession;
         let mut agent = crate::NullAgent;
         let mut ctx = make_ctx(&mut sink, &mut debug, &mut messages, &session, &mut agent);
-        let out = GoalCommand.handle(&mut ctx, "").await.unwrap();
-        assert!(matches!(out, CommandOutput::Message(_)));
+        let result = GoalCommand.handle(&mut ctx, "").await;
+        assert!(result.is_err());
     }
 }
