+++ crates/zeph-commands/src/handlers/misc.rs
@@ -121,81 +121,8 @@ impl CommandHandler<CommandContext<'_>> for ImageCommand {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::context::CommandContext;
+    use crate::handlers::test_helpers::{MockDebug, MockMessages, MockSession, make_ctx};
     use crate::sink::NullSink;
-    use crate::traits::debug::DebugAccess;
-    use crate::traits::messages::MessageAccess;
-    use crate::traits::session::SessionAccess;
-    use std::future::Future;
-    use std::pin::Pin;
-
-    struct MockDebug;
-    impl DebugAccess for MockDebug {
-        fn log_status(&self) -> String {
-            String::new()
-        }
-        fn read_log_tail<'a>(
-            &'a self,
-            _n: usize,
-        ) -> Pin<Box<dyn Future<Output = Option<String>> + Send + 'a>> {
-            Box::pin(async { None })
-        }
-        fn scrub(&self, text: &str) -> String {
-            text.to_owned()
-        }
-        fn dump_status(&self) -> Option<String> {
-            None
-        }
-        fn dump_format_name(&self) -> String {
-            String::new()
-        }
-        fn enable_dump(&mut self, _dir: &str) -> Result<String, CommandError> {
-            Ok(String::new())
-        }
-        fn set_dump_format(&mut self, _name: &str) -> Result<(), CommandError> {
-            Ok(())
-        }
-    }
-
-    struct MockMessages;
-    impl MessageAccess for MockMessages {
-        fn clear_history(&mut self) {}
-        fn queue_len(&self) -> usize {
-            0
-        }
-        fn drain_queue(&mut self) -> usize {
-            0
-        }
-        fn notify_queue_count<'a>(
-            &'a mut self,
-            _count: usize,
-        ) -> Pin<Box<dyn Future<Output = ()> + Send + 'a>> {
-            Box::pin(async {})
-        }
-    }
-
-    struct MockSession;
-    impl SessionAccess for MockSession {
-        fn supports_exit(&self) -> bool {
-            false
-        }
-    }
-
-    fn make_ctx<'a>(
-        sink: &'a mut NullSink,
-        debug: &'a mut MockDebug,
-        messages: &'a mut MockMessages,
-        session: &'a MockSession,
-        agent: &'a mut crate::NullAgent,
-    ) -> CommandContext<'a> {
-        CommandContext {
-            sink,
-            debug,
-            messages,
-            session: session as &dyn SessionAccess,
-            agent,
-        }
-    }
 
     #[test]
     fn cache_stats_name_and_description() {
